deleting-and-renaming-archives-and-fixing-codes
