+++ loan-microservice/src/main/java/com/library/loanMicroservice/controller/BookController.java
@@ -1,17 +1,23 @@
 package com.library.loanMicroservice.controller;
 
 import com.library.loanMicroservice.service.BookService;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/books")
 public class BookController {
 
+    @Autowired
     private final BookService bookService;
 
     public BookController(BookService bookService) {
         this.bookService = bookService;
+    }
+
+    @DeleteMapping("{id}")
+    public ResponseEntity<String> delete(@PathVariable("id") Integer id) {
+        return ResponseEntity.ok().body(this.bookService.deleteById(id));
     }
 }
