+++ loan-microservice/src/main/java/com/library/loanMicroservice/controller/GenreController.java
@@ -1,5 +1,6 @@
 package com.library.loanMicroservice.controller;
 
+import com.library.loanMicroservice.service.GenreService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -7,18 +8,15 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.library.loanMicroservice.service.AuthorService;
-
 @RestController
-@RequestMapping("/authors")
-public class AuthorsController {
+@RequestMapping("/genres")
+public class GenreController {
 
     @Autowired
-    private AuthorService authorService;
-
+    private GenreService genreService;
 
     @DeleteMapping("{id}")
     public ResponseEntity<String> delete(@PathVariable("id") Integer id) {
-        return ResponseEntity.ok().body(authorService.deleteById(id));
+        return ResponseEntity.ok().body(genreService.deleteById(id));
     }
 }
