--- loan-microservice/src/main/java/com/library/loanMicroservice/model/Category.java
@@ -1,35 +1,0 @@
-package com.library.loanMicroservice.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(nullable = false)
-    private String CategoryName;
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public void setCategoryName(String categoryName) {
-        CategoryName = categoryName;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public String getCategoryName() {
-        return CategoryName;
-    }
-    
-}
