--- loan-microservice/src/main/java/com/library/loanMicroservice/repository/BooksRepository.java
@@ -1,9 +1,0 @@
-package com.library.loanMicroservice.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.library.loanMicroservice.model.Books;
-
-public interface BooksRepository extends JpaRepository<Books, Integer> {
-    
-}
