--- loan-microservice/src/main/java/com/library/loanMicroservice/repository/CategoryRepository.java
@@ -1,9 +1,0 @@
-package com.library.loanMicroservice.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.library.loanMicroservice.model.Category;
-
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    
-}
