+++ loan-microservice/src/main/java/com/library/loanMicroservice/service/AuthorService.java
@@ -2,12 +2,12 @@ package com.library.loanMicroservice.service;
 
 import java.util.Optional;
 
+import com.library.loanMicroservice.model.Author;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.library.loanMicroservice.model.Authors;
 import com.library.loanMicroservice.repository.AuthorRepository;
 
 @Service
@@ -17,8 +17,8 @@ public class AuthorService {
     private AuthorRepository authorRepository;
 
     public String deleteById(Integer id) {
-        Optional<Authors> authors = authorRepository.findById(id);
-        if (!authors.isPresent()) {
+        Optional<Author> author = authorRepository.findById(Long.valueOf(id));
+        if (!author.isPresent()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Autor não encontrado.");
         }
         return "Autor deletado.";
